Fixed problem with real_multinomial casting answer to integer.

+++ src/utils/algebra/algebra_utils.f90
@@ -640,7 +640,7 @@ function real_multinomial(top,bottom) result(output)
   
   integer, intent(in) :: top
   integer, intent(in) :: bottom(:)
-  integer             :: output
+  real(dp)            :: output
   
   if (top<0) then
     call err()
